Rename ArgumentValue <-> Argument
