+++ test/hotspot/TestingTestFramework/src/compiler/valhalla/framework/ArgumentValue.java
@@ -1,13 +1,200 @@
 package compiler.valhalla.framework;
 
-public enum ArgumentValue {
-    DEFAULT,
-    NUMBER_MINUS_42,
-    NUMBER_42,
-    BOOLEAN_TOGGLE_FIRST_FALSE,
-    BOOLEAN_TOGGLE_FIRST_TRUE,
-    TRUE,
-    FALSE,
-    RANDOM_ONCE,
-    RANDOM_EACH
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Method;
+import java.lang.reflect.Parameter;
+import java.util.Random;
+
+class ArgumentValue {
+    private static final Random random = new Random();
+
+    final private Object argumentValue;
+    final private boolean isToggleBoolean;
+    final private boolean isRandomEach;
+    final private boolean isFixedRandom;
+    final private Class<?> randomClass;
+    private boolean previousBoolean;
+
+    private ArgumentValue(Object argumentValue, Boolean booleanToggle, boolean isFixedRandom) {
+        this.argumentValue = argumentValue;
+        this.isToggleBoolean = booleanToggle != null;
+        this.previousBoolean = isToggleBoolean && !booleanToggle;
+        this.isRandomEach = false;
+        this.randomClass = null;
+        this.isFixedRandom = isFixedRandom;
+    }
+
+    private ArgumentValue(Object argumentValue, Class<?> randomClass) {
+        this.argumentValue = argumentValue;
+        this.isToggleBoolean = false;
+        this.isRandomEach = true;
+        this.randomClass = randomClass;
+        this.isFixedRandom = false;
+    }
+
+    /**
+     * Return all arguments for the @Arguments annotation.
+     * @param m The @Test method
+     * @return  Return array with Argument objects for each specified argument in the @Arguments annotation of m.
+     *          Return null if method has no @Arguments annotation.
+     */
+    public static ArgumentValue[] getArguments(Method m) {
+        Arguments argumentsAnno = m.getAnnotation(Arguments.class);
+        if (argumentsAnno == null) {
+            return null;
+        }
+        Argument[] values = argumentsAnno.value();
+        ArgumentValue[] arguments = new ArgumentValue[values.length];
+        Class<?>[] declaredParameters = m.getParameterTypes();
+        Parameter[] declaredParameterObjects = m.getParameters();
+        TestFormat.check(values.length == declaredParameters.length,
+                         "Number of argument values provided in @Arguments does not match the number of actual arguments in " + m);
+
+        for (int i = 0; i < values.length; i++) {
+            Argument specifiedArg = values[i];
+            Class<?> parameter = declaredParameters[i];
+            Parameter parameterObj = declaredParameterObjects[i];
+            try {
+                switch (specifiedArg) {
+                    case DEFAULT -> arguments[i] = createDefault(parameter);
+                    case NUMBER_42 -> {
+                        TestFormat.check(isNumber(parameter), "Provided invalid NUMBER_42 argument for non-number " + parameterObj + " for " + m);
+                        arguments[i] = create((byte) 42);
+                    }
+                    case NUMBER_MINUS_42 -> {
+                        TestFormat.check(isNumber(parameter), "Provided invalid NUMBER_MINUS_42 argument for non-number " + parameterObj + " for " + m);
+                        arguments[i] = create((byte) -42);
+                    }
+                    case BOOLEAN_TOGGLE_FIRST_FALSE -> {
+                        TestFormat.check(isBoolean(parameter), "Provided invalid BOOLEAN_TOGGLE_FIRST_FALSE argument for non-boolean " + parameterObj + " for " + m);
+                        arguments[i] = createToggleBoolean(false);
+                    }
+                    case BOOLEAN_TOGGLE_FIRST_TRUE -> {
+                        TestFormat.check(ArgumentValue.isBoolean(parameter), "Provided invalid BOOLEAN_TOGGLE_FIRST_TRUE argument for non-boolean " + parameterObj + " for " + m);
+                        arguments[i] = createToggleBoolean(true);
+                    }
+                    case TRUE -> {
+                        TestFormat.check(ArgumentValue.isBoolean(parameter), "Provided invalid TRUE argument for non-boolean " + parameterObj + " for " + m);
+                        arguments[i] = create(true);
+                    }
+                    case FALSE -> {
+                        TestFormat.check(ArgumentValue.isBoolean(parameter), "Provided invalid FALSE argument for non-boolean " + parameterObj + " for " + m);
+                        arguments[i] = create(false);
+                    }
+                    case RANDOM_ONCE -> {
+                        TestFormat.check(isPrimitiveType(parameter), "Provided invalid RANDOM_ONCE argument for non-primitive type " + parameterObj + " for " + m);
+                        arguments[i] = createRandom(parameter);
+                    }
+                    case RANDOM_EACH -> {
+                        TestFormat.check(isPrimitiveType(parameter), "Provided invalid RANDOM_ONCE argument for non-primitive type " + parameterObj + " for " + m);
+                        arguments[i] = createRandomEach(parameter);
+                    }
+                }
+            } catch (TestFormatException e) {
+                // Catch and continue to check arguments.
+            }
+        }
+        return arguments;
+    }
+
+    private static ArgumentValue create(Object argumentValue) {
+        return new ArgumentValue(argumentValue, null, false);
+    }
+
+    private static ArgumentValue createDefault(Class<?> c) {
+        if (ArgumentValue.isNumber(c)) {
+            return ArgumentValue.create((byte)0);
+        } else if (ArgumentValue.isChar(c)) {
+            return ArgumentValue.create('\u0000');
+        } else if (ArgumentValue.isBoolean(c)) {
+            return ArgumentValue.create(false);
+        } else {
+            // Object
+            try {
+                Constructor<?> constructor = c.getDeclaredConstructor();
+                constructor.setAccessible(true);
+                return ArgumentValue.create(constructor.newInstance());
+            } catch (Exception e) {
+                TestFormat.fail("Cannot create new default instance of " + c + ": " + e.getCause());
+                return null;
+            }
+        }
+    }
+
+    private static ArgumentValue createRandom(Class<?> c) {
+        return new ArgumentValue(getRandom(c), null, true);
+    }
+    private static ArgumentValue createToggleBoolean(boolean firstBoolean) {
+        return new ArgumentValue(null, firstBoolean, false);
+    }
+
+    private static ArgumentValue createRandomEach(Class<?> c) {
+        return new ArgumentValue(null, c);
+    }
+
+    public boolean isFixedRandom() {
+        return isFixedRandom;
+    }
+
+    public Object getArgument() {
+        if (isToggleBoolean) {
+            previousBoolean = !previousBoolean;
+            return previousBoolean;
+        } else if (isRandomEach) {
+            return getRandom(randomClass);
+        } else {
+            return argumentValue;
+        }
+    }
+
+    private static boolean isPrimitiveType(Class<?> c) {
+        return isNumber(c) || isBoolean(c) || isChar(c);
+    }
+
+    private static boolean isBoolean(Class<?> c) {
+        return c.equals(boolean.class);
+    }
+
+    private static boolean isChar(Class<?> c) {
+        return c.equals(char.class);
+    }
+    private static boolean isNumber(Class<?> c) {
+        return isIntNumber(c) || isFloatNumber(c);
+    }
+
+    private static boolean isIntNumber(Class<?> c) {
+        return c.equals(byte.class)
+                || c.equals(short.class)
+                || c.equals(int.class)
+                || c.equals(long.class);
+    }
+
+    public static boolean isFloatNumber(Class<?> c) {
+        return c.equals(float.class) || c.equals(double.class);
+    }
+
+    private static Object getRandom(Class<?> c) {
+        if (isBoolean(c)) {
+            return random.nextBoolean();
+        } else if (c.equals(byte.class)) {
+            return (byte)random.nextInt(256);
+        } else if (isChar(c)) {
+            return (char)random.nextInt(65536);
+        }  else if (c.equals(short.class)) {
+            return (short)random.nextInt(65536);
+        } else if (c.equals(int.class)) {
+            return random.nextInt();
+        } else if (c.equals(long.class)) {
+            return random.nextLong();
+        } else if (c.equals(float.class)) {
+            // Get number between 0 and 1000.
+            return random.nextFloat() * 1000;
+        } else if (c.equals(double.class)) {
+            // Get number between 0 and 1000.
+            return random.nextDouble() * 1000;
+        } else {
+            TestFormat.fail("Cannot generate random value for non-primitive type");
+            return null;
+        }
+    }
 }
