+++ test/hotspot/TestingTestFramework/src/compiler/valhalla/framework/tests/TestBasics.java
@@ -137,7 +137,7 @@ public class TestBasics {
     // Base test, with arguments, directly invoked.
     // Specify the argument values with @Arguments
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void byteDefaultArgument(byte x) {
         executed[4]++;
         if (x != 0) {
@@ -146,7 +146,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void shortDefaultArgument(short x) {
         executed[5]++;
         if (x != 0) {
@@ -155,7 +155,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void intDefaultArgument(int x) {
         executed[6]++;
         if (x != 0) {
@@ -164,7 +164,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void longDefaultArgument(long x) {
         executed[7]++;
         if (x != 0L) {
@@ -173,7 +173,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void floatDefaultArgument(float x) {
         executed[8]++;
         if (x != 0.0f) {
@@ -182,7 +182,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void doubleDefaultArgument(double x) {
         executed[9]++;
         if (x != 0.0f) {
@@ -191,7 +191,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void charDefaultArgument(char x) {
         executed[10]++;
         if (x != '\u0000') {
@@ -200,7 +200,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void booleanDefaultArgument(boolean x) {
         executed[11]++;
         if (x) {
@@ -209,7 +209,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void stringObjectDefaultArgument(String x) {
         executed[12]++;
         if (x == null || x.length() != 0) {
@@ -218,7 +218,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void defaultObjectDefaultArgument(DefaultObject x) {
         executed[13]++;
         if (x == null || x.i != 4) {
@@ -227,7 +227,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_42)
+    @Arguments(Argument.NUMBER_42)
     public void byte42(byte x) {
         executed[14]++;
         if (x != 42) {
@@ -236,7 +236,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_42)
+    @Arguments(Argument.NUMBER_42)
     public void short42(short x) {
         executed[15]++;
         if (x != 42) {
@@ -245,7 +245,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_42)
+    @Arguments(Argument.NUMBER_42)
     public void int42(int x) {
         executed[16]++;
         if (x != 42) {
@@ -254,7 +254,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_42)
+    @Arguments(Argument.NUMBER_42)
     public void long42(long x) {
         executed[17]++;
         if (x != 42) {
@@ -263,7 +263,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_42)
+    @Arguments(Argument.NUMBER_42)
     public void float42(float x) {
         executed[18]++;
         if (x != 42.0) {
@@ -272,7 +272,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_42)
+    @Arguments(Argument.NUMBER_42)
     public void double42(double x) {
         executed[19]++;
         if (x != 42.0) {
@@ -281,7 +281,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.FALSE)
+    @Arguments(Argument.FALSE)
     public void booleanFalse(boolean x) {
         executed[20]++;
         if (x) {
@@ -290,7 +290,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.TRUE)
+    @Arguments(Argument.TRUE)
     public void booleanTrue(boolean x) {
         executed[21]++;
         if (!x) {
@@ -299,37 +299,37 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomByte(byte x) {
         executed[22]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomShort(short x) {
         executed[23]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomInt(int x) {
         executed[24]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomLong(long x) {
         executed[25]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomFloat(float x) {
         executed[26]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomDouble(double x) {
         executed[27]++;
     }
@@ -340,13 +340,13 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_ONCE)
+    @Arguments(Argument.RANDOM_ONCE)
     public void randomBoolean(boolean x) {
         executed[28]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.BOOLEAN_TOGGLE_FIRST_FALSE)
+    @Arguments(Argument.BOOLEAN_TOGGLE_FIRST_FALSE)
     public void booleanToggleFirstFalse(boolean x) {
         if (executed[29] == 0) {
             // First invocation
@@ -361,56 +361,56 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachByte(byte x) {
         checkNonFloatingRandomNumber(x, executed[30]);
         executed[30]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachShort(short x) {
         checkNonFloatingRandomNumber(x, executed[31]);
         executed[31]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachInt(int x) {
         checkNonFloatingRandomNumber(x, executed[32]);
         executed[32]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachLong(long x) {
         checkNonFloatingRandomNumber(x, executed[33]);
         executed[33]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachChar(char x) {
         checkNonFloatingRandomNumber(x, executed[34]);
         executed[34]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachFloat(float x) {
         checkFloatingRandomNumber(x, executed[35]);
         executed[35]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachDouble(double x) {
         checkFloatingRandomNumber(x, executed[36]);
         executed[36]++;
     }
 
     @Test
-    @Arguments(ArgumentValue.RANDOM_EACH)
+    @Arguments(Argument.RANDOM_EACH)
     public void randomEachBoolean(boolean x) {
         checkRandomBoolean(x, executed[37]);
         executed[37]++;
@@ -463,7 +463,7 @@ public class TestBasics {
 
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_MINUS_42)
+    @Arguments(Argument.NUMBER_MINUS_42)
     public void byteMinus42(byte x) {
         executed[38]++;
         if (x != -42) {
@@ -472,7 +472,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_MINUS_42)
+    @Arguments(Argument.NUMBER_MINUS_42)
     public void shortMinus42(short x) {
         executed[39]++;
         if (x != -42) {
@@ -481,7 +481,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_MINUS_42)
+    @Arguments(Argument.NUMBER_MINUS_42)
     public void intMinus42(int x) {
         executed[40]++;
         if (x != -42) {
@@ -490,7 +490,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_MINUS_42)
+    @Arguments(Argument.NUMBER_MINUS_42)
     public void longMinus42(long x) {
         executed[41]++;
         if (x != -42) {
@@ -499,7 +499,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_MINUS_42)
+    @Arguments(Argument.NUMBER_MINUS_42)
     public void floatMinus42(float x) {
         executed[42]++;
         if (x != -42.0) {
@@ -508,7 +508,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments(ArgumentValue.NUMBER_MINUS_42)
+    @Arguments(Argument.NUMBER_MINUS_42)
     public void doubleMinus42(double x) {
         executed[43]++;
         if (x != -42.0) {
@@ -517,7 +517,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.DEFAULT, ArgumentValue.DEFAULT})
+    @Arguments({Argument.DEFAULT, Argument.DEFAULT})
     public void twoArgsDefault1(byte x, short y) {
         executed[44]++;
         if (x != 0 || y != 0) {
@@ -526,7 +526,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.DEFAULT, ArgumentValue.DEFAULT})
+    @Arguments({Argument.DEFAULT, Argument.DEFAULT})
     public void twoArgsDefault2(int x, short y) {
         executed[45]++;
         if (x != 0 || y != 0) {
@@ -535,7 +535,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.DEFAULT, ArgumentValue.DEFAULT})
+    @Arguments({Argument.DEFAULT, Argument.DEFAULT})
     public void twoArgsDefault3(short x, long y) {
         executed[46]++;
         if (x != 0 || y != 0) {
@@ -544,7 +544,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.DEFAULT, ArgumentValue.DEFAULT})
+    @Arguments({Argument.DEFAULT, Argument.DEFAULT})
     public void twoArgsDefault4(float x, boolean y) {
         executed[47]++;
         if (x != 0.0 || y) {
@@ -553,7 +553,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.DEFAULT, ArgumentValue.DEFAULT})
+    @Arguments({Argument.DEFAULT, Argument.DEFAULT})
     public void twoArgsDefault5(boolean x, char y) {
         executed[48]++;
         if (x || y != '\u0000') {
@@ -562,7 +562,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.DEFAULT, ArgumentValue.DEFAULT})
+    @Arguments({Argument.DEFAULT, Argument.DEFAULT})
     public void twoArgsDefault6(char x, byte y) {
         executed[49]++;
         if (x != '\u0000' || y != 0) {
@@ -571,16 +571,16 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE})
+    @Arguments({Argument.RANDOM_ONCE, Argument.RANDOM_ONCE})
     public void twoArgsRandomOnce(char x, byte y) {
         executed[50]++;
     }
 
     @Test
-    @Arguments({ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE})
+    @Arguments({Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_ONCE, Argument.RANDOM_ONCE})
     public void checkRandomOnceDifferentArgs(int a, int b, int c, int d, int e, int f, int g, int h) {
         if (Stream.of(a, b, c, d, e, f, g, h).allMatch(i -> i == a)) {
             throw new RuntimeException("RANDOM_ONCE does not produce random values for different arguments");
@@ -589,36 +589,36 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE})
+    @Arguments({Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_ONCE, Argument.RANDOM_ONCE})
     public void checkMixedRandoms1(byte a, short b, int c, long d, char e, boolean f, float g, double h) {
         executed[52]++;
     }
 
     @Test
-    @Arguments({ArgumentValue.RANDOM_EACH, ArgumentValue.RANDOM_EACH,
-                ArgumentValue.RANDOM_EACH, ArgumentValue.RANDOM_EACH,
-                ArgumentValue.RANDOM_EACH, ArgumentValue.RANDOM_EACH,
-                ArgumentValue.RANDOM_EACH, ArgumentValue.RANDOM_EACH})
+    @Arguments({Argument.RANDOM_EACH, Argument.RANDOM_EACH,
+                Argument.RANDOM_EACH, Argument.RANDOM_EACH,
+                Argument.RANDOM_EACH, Argument.RANDOM_EACH,
+                Argument.RANDOM_EACH, Argument.RANDOM_EACH})
     public void checkMixedRandoms2(byte a, short b, int c, long d, char e, boolean f, float g, double h) {
         executed[53]++;
     }
 
     @Test
-    @Arguments({ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_ONCE,
-                ArgumentValue.RANDOM_EACH, ArgumentValue.RANDOM_EACH,
-                ArgumentValue.RANDOM_ONCE, ArgumentValue.RANDOM_EACH,
-                ArgumentValue.RANDOM_EACH, ArgumentValue.RANDOM_ONCE})
+    @Arguments({Argument.RANDOM_ONCE, Argument.RANDOM_ONCE,
+                Argument.RANDOM_EACH, Argument.RANDOM_EACH,
+                Argument.RANDOM_ONCE, Argument.RANDOM_EACH,
+                Argument.RANDOM_EACH, Argument.RANDOM_ONCE})
     public void checkMixedRandoms3(byte a, short b, int c, long d, char e, boolean f, float g, double h) {
         executed[54]++;
     }
 
     @Test
-    @Arguments({ArgumentValue.NUMBER_42, ArgumentValue.NUMBER_42,
-                ArgumentValue.NUMBER_42, ArgumentValue.NUMBER_42,
-                ArgumentValue.NUMBER_42, ArgumentValue.NUMBER_42})
+    @Arguments({Argument.NUMBER_42, Argument.NUMBER_42,
+                Argument.NUMBER_42, Argument.NUMBER_42,
+                Argument.NUMBER_42, Argument.NUMBER_42})
     public void check42Mix1(byte a, short b, int c, long d, float e, double f) {
         if (a != 42 || b != 42 || c != 42 || d != 42 || e != 42.0 || f != 42.0) {
             throw new RuntimeException("Must all be 42");
@@ -627,9 +627,9 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.NUMBER_MINUS_42, ArgumentValue.NUMBER_MINUS_42,
-                ArgumentValue.NUMBER_MINUS_42, ArgumentValue.NUMBER_MINUS_42,
-                ArgumentValue.NUMBER_MINUS_42, ArgumentValue.NUMBER_MINUS_42})
+    @Arguments({Argument.NUMBER_MINUS_42, Argument.NUMBER_MINUS_42,
+                Argument.NUMBER_MINUS_42, Argument.NUMBER_MINUS_42,
+                Argument.NUMBER_MINUS_42, Argument.NUMBER_MINUS_42})
     public void check42Mix2(byte a, short b, int c, long d, float e, double f) {
         if (a != -42 || b != -42 || c != -42 || d != -42 || e != -42.0 || f != -42.0) {
             throw new RuntimeException("Must all be -42");
@@ -638,9 +638,9 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.NUMBER_MINUS_42, ArgumentValue.NUMBER_42,
-                ArgumentValue.NUMBER_MINUS_42, ArgumentValue.NUMBER_MINUS_42,
-                ArgumentValue.NUMBER_42, ArgumentValue.NUMBER_MINUS_42})
+    @Arguments({Argument.NUMBER_MINUS_42, Argument.NUMBER_42,
+                Argument.NUMBER_MINUS_42, Argument.NUMBER_MINUS_42,
+                Argument.NUMBER_42, Argument.NUMBER_MINUS_42})
     public void check42Mix3(byte a, short b, int c, long d, float e, double f) {
         if (a != -42 || b != 42 || c != -42 || d != -42 || e != 42.0 || f != -42.0) {
             throw new RuntimeException("Do not match the right 42 version");
@@ -650,7 +650,7 @@ public class TestBasics {
 
 
     @Test
-    @Arguments(ArgumentValue.BOOLEAN_TOGGLE_FIRST_TRUE)
+    @Arguments(Argument.BOOLEAN_TOGGLE_FIRST_TRUE)
     public void booleanToggleFirstTrue(boolean x) {
         if (executed[58] == 0) {
             // First invocation
@@ -665,7 +665,7 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.BOOLEAN_TOGGLE_FIRST_FALSE, ArgumentValue.BOOLEAN_TOGGLE_FIRST_TRUE})
+    @Arguments({Argument.BOOLEAN_TOGGLE_FIRST_FALSE, Argument.BOOLEAN_TOGGLE_FIRST_TRUE})
     public void checkTwoToggles(boolean b1, boolean b2) {
         if (executed[59] == 0) {
             // First invocation
@@ -682,8 +682,8 @@ public class TestBasics {
     }
 
     @Test
-    @Arguments({ArgumentValue.BOOLEAN_TOGGLE_FIRST_FALSE, ArgumentValue.FALSE,
-                ArgumentValue.TRUE, ArgumentValue.BOOLEAN_TOGGLE_FIRST_TRUE})
+    @Arguments({Argument.BOOLEAN_TOGGLE_FIRST_FALSE, Argument.FALSE,
+                Argument.TRUE, Argument.BOOLEAN_TOGGLE_FIRST_TRUE})
     public void booleanMix(boolean b1, boolean b2, boolean b3, boolean b4) {
         if (executed[60] == 0) {
             // First invocation
