+++ test/hotspot/TestingTestFramework/src/compiler/valhalla/framework/tests/TestPackagePrivate.java
@@ -2,11 +2,6 @@ package compiler.valhalla.framework.tests;
 
 import compiler.valhalla.framework.*;
 
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-import java.util.Arrays;
-import java.util.stream.Stream;
-
 public class TestPackagePrivate {
     public static void main(String[] args) {
         TestFramework.run(PackagePrivate.class);
@@ -19,7 +14,7 @@ class PackagePrivate {
     }
 
     @Test
-    @Arguments(ArgumentValue.DEFAULT)
+    @Arguments(Argument.DEFAULT)
     public void test2(int x) {
     }
 }
